requirements.txt
stmpy
paho-mqtt
Pillow
Flask
flask-cors
flask-sock
requests
websocket-client
